// src/lib/weebdex/model/getMangaDemographicItem.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * WeebDex API
 * WeebDex is an ad-free, scanlator-first online manga reader providing high-quality, zero-compression images. <br /><br />
Contact: [delta@weebdex.org](mailto:delta@weebdex.org) or join our [Discord Server](https://discord.gg/syp3a6Eucq).<br /><br />
All API endpoints have a global rate limit of 5 requests per second per IP. <br />
To avoid future issues, include the Origin: https://weebdex.org and Referer: https://weebdex.org/ headers when making API requests.<br/>
 * OpenAPI spec version: 1.2.0
 */

export type GetMangaDemographicItem =
  (typeof GetMangaDemographicItem)[keyof typeof GetMangaDemographicItem];

export const GetMangaDemographicItem = {
  shounen: "shounen",
  shoujo: "shoujo",
  josei: "josei",
  seinen: "seinen",
  none: "none",
} as const;